SQL database based mass spectrometry data backend

The MsBackendSqlDb is an MsBackend backend that keeps all metadata and spectra data (m/z and intensity values) in a single on-disk database table. This object has a very low memory footprint because only primary keys of the database table are stored in memory, within the object hence making it ideal for representing very large mass spectrometry data sets on computers with limited resources.

MsBackendSqlDb()

# S4 method for MsBackendSqlDb
backendInitialize(object, dbcon, files = character(), dbtable = "msdata")

Arguments

object
a MsBackendSqlDb object.

dbcon
a DBIConnection object to connect to the database.

files
character with the file names from which data should be imported. Only required if the database to which dbcon connects does not already contain the data.

dbtable
character(1) the name of the database table with the data. Defaults to dbtable = "msdata".

Slots

dbtable

A character(1) with the name of the database table (or view) containing the data.

dbcon

A DBIConnection with the connection to the database.

modCount

An integer() that keeps track of each database writing cycle which would invalidate objects pointing to the same database. This number has to match between the database and the object.

rows

An integer() with the indices (primary keys) of the data.

columns

A character() containing the names of the columns stored in the database.

`query`

A DBIResults object containing SQL query against the backend.

Creating an MsBackendSqlDb instance

Like all MsBackend() objects, MsBackendSqlDb have to be initialized with the backendInitialize() method that either fills the object with data from mass spectrometry files provided with the files parameter or checks if the pre-generated database contains all required data. A connection to a database (with write access) has to be provided with the dbcon parameter which expects a DBIConnection returned by a call to dbConnect(). To fill the database with MS data files (usually mzML files, but all files types supported by the mzR package are allowed), these have to be provided with the files parameter. Parameter dbtable can be used to optionally specify the database table name that contains, or should contain the data.

Implementation notes

The MsBackendSqlDb defines the following slots which should not be accessed or changed by the user.

Author

Johannes Rainer

Examples

## Initialize an MsBackendSqlDb filling it with data from mzML files
library(msdata)
library(RSQLite)
con <- dbConnect(SQLite(), tempfile())
fls <- dir(system.file("sciex", package = "msdata"), full.names = TRUE,
    pattern = "mzML$")
msb <- backendInitialize(MsBackendSqlDb(), dbcon = con, files = fls)